+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog #
 
-## Upcoming changes
+## Upcoming changes (v1.3.0)
 
 - Better error reporting
 - Improved charts.
+- Add healthcheck/keepalive monitor.
+- Restyle monitor detail page.
+- Add log visualization page for healthcheck monitors.
+- Add the ability to export status history as CSV and JSON.
+- Add the ability to export logs as CSV and JSON.
+- Fix issues on export/import.
 
 
 ## v1.2.0